Sampling time checked
The maximum cycling time of unloaded program is 10400 us: 100-200 us is a basic value, but 1e4 appears time to time for unknown reason (Servo lib inner procedures maybe?). Before the limitation with this aspect is occured it is reasonable to proceed with explicitly:
#define TIME_STAMP 20000 // us
to keep the future model and the device consistent.